validate notification email from standard notification
