+++ aldryn_forms/contrib/email_notifications/cms_plugins.py
@@ -8,9 +8,9 @@ from django.utils.translation import ugettext_lazy as _
 from cms.plugin_pool import plugin_pool
 
 from aldryn_forms.cms_plugins import FormPlugin
+from aldryn_forms.validators import is_valid_recipient
 
 from .models import EmailNotification, EmailNotificationFormPlugin
-from .validators import is_valid_recipient
 
 
 logger = logging.getLogger(__name__)
